chore: extract tessen into a createTessen helper
